+++ compiler-tests/src/test/data/box/bindingcontainers/SimpleContainersWithHintsWork.kt
@@ -1,0 +1,18 @@
+// https://github.com/ZacSweers/metro/issues/733
+@SingleIn(AppScope::class)
+@DependencyGraph(AppScope::class, isExtendable = true)
+interface AppGraph {
+  val text: String
+}
+
+@ContributesTo(AppScope::class)
+@BindingContainer
+object SampleDependency {
+  @Provides fun provideText(): String = "Hello, Metro!"
+}
+
+fun box(): String {
+  val graph = createGraph<AppGraph>()
+  assertEquals("Hello, Metro!", graph.text)
+  return "OK"
+}